requirements.txt
aiofiles==0.6.0
captcha==0.3
click==7.1.2
docxtpl==0.11.5
fastapi==0.64.0
h11==0.12.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
Pillow==8.2.0
pydantic==1.8.1
python-docx==0.8.10
python-dotenv==0.17.1
randomuser==1.6
six==1.16.0
starlette==0.13.6
typing-extensions==3.10.0.0
uvicorn==0.13.4
